Refactor InliningState (#39881)
We recently dropped the method table from this and we can get rid
of one of the cache references now too, since we're now passing that
info in the stmtinfo. Also factor out the inlining policy into a
separate option to the inliner to make it easier for downstream
projects like Diffractor and Cthulhu to override it without playing
cache games (which is now harder, since one of the caches is gone).